FlashRep/src/main/java/flashrep/flashrep/logic/Flashcard.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package flashrep.flashrep.logic;

/**
 *
 * @author Raine Rantanen
 */
public class Flashcard {

    private String question;
    private String answer;

    public Flashcard(String question, String answer) {
        this.question = question;
        this.answer = answer;
    }

    public String getAnswer() {
        return answer;
    }

    public String getQuestion() {
        return question;
    }

    public void setAnswer(String answer) {
        this.answer = answer;
    }

    public void setQuestion(String question) {
        this.question = question;
    }

    public void swap() {
        String question = this.question;
        String answer = this.answer;

        this.question = answer;
        this.answer = question;
    }
}
